pyproject.toml
[tool.poetry]
name = "py_omop2neo4j_lpg"
version = "0.2.0"
description = "A Python package to orchestrate the migration of OMOP vocabulary tables from PostgreSQL to a Labeled Property Graph (LPG) in Neo4j."
authors = ["Gowtham Adamane Rao <rao@ohdsi.org>"]
readme = "README.md"
license = "Prosperity Public License 3.0.0"
homepage = "https://github.com/gowthamrao/py-omop2neo4j-lpg"
repository = "https://github.com/gowthamrao/py-omop2neo4j-lpg"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Database",
]
packages = [{include = "py_omop2neo4j_lpg", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11.0"
neo4j = "*"
pandas = "*"
numpy = "*"
click = "*"
pydantic-settings = "*"
python-dotenv = "*"

[tool.poetry.group.test.dependencies]
pytest = "*"
psycopg2-binary = "*"
pytest-docker = "*"
pytest-dotenv = "*"
pytest-cov = "*"
ruff = "*"
mypy = "*"
pre-commit = "*"

[tool.poetry.scripts]
py_omop2neo4j_lpg = "py_omop2neo4j_lpg.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (require docker)",
    "unit: marks tests as unit tests",
]
env_files = [
    ".env.test"
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
